requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
websockets>=12.0
solders>=0.21.0
solana>=0.34.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
